+++ src/dimmer_protocol.h
@@ -32,7 +32,6 @@
 #define DIMMER_REGISTER_MAX_TEMP            (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, max_temp))
 #define DIMMER_REGISTER_FADE_IN_TIME        (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, fade_in_time))
 #define DIMMER_REGISTER_TEMP_CHECK_INT      (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, temp_check_interval))
-// #define DIMMER_REGISTER_LC_FACTOR           (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, linear_correction_factor))
 #define DIMMER_REGISTER_ZC_DELAY_TICKS      (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, zero_crossing_delay_ticks))
 #define DIMMER_REGISTER_MIN_ON_TIME_TICKS   (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, minimum_on_time_ticks))
 #define DIMMER_REGISTER_MIN_OFF_TIME_TICKS  (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, minimum_off_time_ticks))
@@ -40,8 +39,6 @@
 #define DIMMER_REGISTER_INT_TEMP_OFS        (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, int_temp_offset))
 #define DIMMER_REGISTER_METRICS_INT         (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, report_metrics_max_interval))
 #define DIMMER_REGISTER_VERSION             0xb9
-//#define DIMMER_REGISTER_VERSION             (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, version))
-//#define DIMMER_REGISTER_VERSION             (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, version))
 #define DIMMER_REGISTER_RANGE_BEGIN         (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, range_begin))
 #define DIMMER_REGISTER_RANGE_END           (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, range_end))
 #define DIMMER_REGISTER_SWITCH_ON_MIN_TIME  (DIMMER_REGISTER_START_ADDR + offsetof(register_mem_t, cfg) + offsetof(register_mem_cfg_t, switch_on_minimum_ticks))
@@ -67,7 +64,7 @@
 #define DIMMER_COMMAND_READ_AC_FREQUENCY    0x23
 #define DIMMER_COMMAND_WRITE_EEPROM         0x50
 #define DIMMER_COMMAND_RESTORE_FS           0x51
-#define DIMMER_COMMAND_READ_TIMINGS         0x52
+#define DIMMER_COMMAND_GET_TIMER_TICKS      0x52
 #define DIMMER_COMMAND_PRINT_INFO           0x53
 #define DIMMER_COMMAND_FORCE_TEMP_CHECK     0x54
 #define DIMMER_COMMAND_PRINT_METRICS        0x55
@@ -90,88 +87,4 @@
 #define DIMMER_OPTIONS_REPORT_TEMP          0x02
 #define DIMMER_OPTIONS_TEMP_ALERT_TRIGGERED 0x04
 
-// DIMMER_COMMAND_READ_TIMINGS
-// all timings are float and read from DIMMER_REGISTER_TEMP
-#define DIMMER_TIMINGS_TMR1_TICKS_PER_US    0x01
-#define DIMMER_TIMINGS_TMR2_TICKS_PER_US    0x02
-#define DIMMER_TIMINGS_ZC_DELAY_IN_US       0x03
-#define DIMMER_TIMINGS_MIN_ON_TIME_IN_US    0x04
-#define DIMMER_TIMINGS_MIN_OFF_TIME_IN_US   0x05
-
 static_assert(DIMMER_REGISTER_VERSION == 0xb9, "Invalid offset");
-
-
-/*
-#define DIMMER_REGISTER_START_ADDR               0x80
-#define DIMMER_REGISTER_FROM_LEVEL               0x80
-#define DIMMER_REGISTER_CHANNEL                  0x82
-#define DIMMER_REGISTER_TO_LEVEL                 0x83
-#define DIMMER_REGISTER_TIME                     0x85
-#define DIMMER_REGISTER_COMMAND                  0x89
-#define DIMMER_REGISTER_READ_LENGTH              0x8a
-#define DIMMER_REGISTER_COMMAND_STATUS           0x8b
-#define DIMMER_REGISTER_CHANNELS_START           0x8c
-#define DIMMER_REGISTER_CH0_LEVEL                0x8c
-#define DIMMER_REGISTER_CH1_LEVEL                0x8e
-#define DIMMER_REGISTER_CH2_LEVEL                0x90
-#define DIMMER_REGISTER_CH3_LEVEL                0x92
-#define DIMMER_REGISTER_CH4_LEVEL                0x94
-#define DIMMER_REGISTER_CH5_LEVEL                0x96
-#define DIMMER_REGISTER_CH6_LEVEL                0x98
-#define DIMMER_REGISTER_CH7_LEVEL                0x9a
-#define DIMMER_REGISTER_CHANNELS_END             0x9c
-#define DIMMER_REGISTER_TEMP                     0x9c
-#define DIMMER_REGISTER_VCC                      0xa0
-#define DIMMER_REGISTER_OPTIONS                  0xa2
-#define DIMMER_REGISTER_MAX_TEMP                 0xa3
-#define DIMMER_REGISTER_FADE_IN_TIME             0xa4
-#define DIMMER_REGISTER_TEMP_CHECK_INT           0xa8
-#define DIMMER_REGISTER_ZC_DELAY_TICKS           0xad
-#define DIMMER_REGISTER_MIN_ON_TIME_TICKS        0xae
-#define DIMMER_REGISTER_MIN_OFF_TIME_TICKS       0xb0
-#define DIMMER_REGISTER_INT_1_1V_REF             0xb2
-#define DIMMER_REGISTER_INT_TEMP_OFS             0xb6
-#define DIMMER_REGISTER_METRICS_INT              0xb8
-#define DIMMER_REGISTER_VERSION                  0xb9
-#define DIMMER_REGISTER_RANGE_BEGIN              0xb9
-#define DIMMER_REGISTER_RANGE_END                0xbb
-#define DIMMER_REGISTER_SWITCH_ON_MIN_TIME       0xbd
-#define DIMMER_REGISTER_SWITCH_ON_COUNT          0xbf
-#define DIMMER_REGISTER_ADDRESS                  0xc3
-#define DIMMER_REGISTER_END_ADDR                 0xc4
-
-#define DIMMER_METRICS_REPORT                    0xf0
-#define DIMMER_TEMPERATURE_ALERT                 0xf1
-#define DIMMER_FADING_COMPLETE                   0xf2
-#define DIMMER_EEPROM_WRITTEN                    0xf3
-#define DIMMER_FREQUENCY_WARNING                 0xf4
-#define DIMMER_CHANNEL_ON_OFF                    0xf5
-
-#define DIMMER_COMMAND_SET_LEVEL                 0x10
-#define DIMMER_COMMAND_FADE                      0x11
-#define DIMMER_COMMAND_READ_CHANNELS             0x12
-#define DIMMER_COMMAND_READ_NTC                  0x20
-#define DIMMER_COMMAND_READ_INT_TEMP             0x21
-#define DIMMER_COMMAND_READ_VCC                  0x22
-#define DIMMER_COMMAND_READ_AC_FREQUENCY         0x23
-#define DIMMER_COMMAND_WRITE_EEPROM              0x50
-#define DIMMER_COMMAND_RESTORE_FS                0x51
-#define DIMMER_COMMAND_READ_TIMINGS              0x52
-#define DIMMER_COMMAND_PRINT_INFO                0x53
-#define DIMMER_COMMAND_FORCE_TEMP_CHECK          0x54
-#define DIMMER_COMMAND_PRINT_METRICS             0x55
-#define DIMMER_COMMAND_ZC_TIMINGS_OUTPUT         0x60
-#define DIMMER_COMMAND_DUMP_MEM                  0xee
-#define DIMMER_COMMAND_STATUS_OK                 0x00
-#define DIMMER_COMMAND_STATUS_ERROR              0xff
-
-#define DIMMER_OPTIONS_RESTORE_LEVEL             0x01
-#define DIMMER_OPTIONS_REPORT_TEMP               0x02
-#define DIMMER_OPTIONS_TEMP_ALERT_TRIGGERED      0x04
-
-#define DIMMER_TIMINGS_TMR1_TICKS_PER_US         0x01
-#define DIMMER_TIMINGS_TMR2_TICKS_PER_US         0x02
-#define DIMMER_TIMINGS_ZC_DELAY_IN_US            0x03
-#define DIMMER_TIMINGS_MIN_ON_TIME_IN_US         0x04
-#define DIMMER_TIMINGS_MIN_OFF_TIME_IN_US        0x05
-*/